#853 refactor training calculation (1st steps)
