+++ src/main/java/core/model/player/Player.java
@@ -9,6 +9,9 @@ import core.model.FactorObject;
 import core.model.FormulaFactors;
 import core.model.HOVerwaltung;
 import core.model.StaffMember;
+import core.model.match.MatchLineupTeam;
+import core.model.match.MatchType;
+import core.model.match.SourceSystem;
 import core.model.match.Weather;
 import core.model.misc.TrainingEvent;
 import core.net.OnlineWorker;
@@ -20,6 +23,8 @@ import core.util.HelperWrapper;
 
 import java.sql.Timestamp;
 import java.util.*;
+
+import static java.lang.Integer.min;
 
 public class Player {
     //~ Class fields -------------------------------------------------------------------------------
@@ -1864,7 +1869,7 @@ public class Player {
      * Calculates training benefit, and updates subskill for the player.
      * Used when there is only 1 week of training to be calculated.
      *
-     * @param originalPlayer - The player to calculate subskills on
+     * @param originalPlayer - The player to calculate subskills on. Status of previous training.
      * @param trainerlevel   - The trainer level
      * @param intensity      - Training intensity
      * @param stamina
@@ -1875,13 +1880,13 @@ public class Player {
                                          int stamina, TrainingPerWeek trainingWeek, List<StaffMember> staff) {
 
         if (this.hasTrainingBlock()) {
-            return;
+            return;    // training of the player is disabled
         }
 
         if (trainingWeek == null)
             return;
 
-        TrainingPerPlayer trForPlayer = TrainingManager.instance().calculateWeeklyTrainingForPlayer(this, trainingWeek);
+        TrainingPerPlayer trForPlayer = this.calculateWeeklyTraining( trainingWeek);
 
         if (trForPlayer == null)
             return;
@@ -1906,6 +1911,73 @@ public class Player {
 
         addExperienceSub(trForPlayer.getExperienceSub());
 
+    }
+
+    /**
+     * Calculate Training for each skill
+     *
+     * @param train preset training weeks
+     *
+     * @return TrainingPerPlayer
+     */
+    public TrainingPerPlayer calculateWeeklyTraining(TrainingPerWeek train) {
+        final int playerID = this.getPlayerID();
+        TrainingPerPlayer output = new TrainingPerPlayer(this);
+        if (train == null || train.getTrainingType() < 0) {
+            return output;
+        }
+
+        WeeklyTrainingType wt = WeeklyTrainingType.instance(train.getTrainingType());
+        if (wt != null) {
+            try {
+                var matches = train.getMatches();
+                int myID = HOVerwaltung.instance().getModel().getBasics().getTeamId();
+                TrainingWeekPlayer tp = new TrainingWeekPlayer(this);
+                int minutes=0;
+                for (var match : matches) {
+                    //Get the MatchLineup by id
+                    MatchLineupTeam mlt = DBManager.instance().getMatchLineupTeam(SourceSystem.HATTRICK.getValue(), match.getMatchID(), myID);
+                    MatchType type = mlt.getMatchType();
+                    boolean walkoverWin = match.getMatchdetails().isWalkoverMatchWin(HOVerwaltung.instance().getModel().getBasics().getYouthTeamId());
+                    if ( type != MatchType.MASTERS) { // MASTERS counts only for experience
+                        tp.addFullTrainingMinutes(mlt.getTrainingMinutesPlayedInSectors(playerID, wt.getFullTrainingSectors(), walkoverWin));
+                        tp.addBonusTrainingMinutes(mlt.getTrainingMinutesPlayedInSectors(playerID, wt.getBonusTrainingSectors(), walkoverWin));
+                        tp.addPartlyTrainingMinutes(mlt.getTrainingMinutesPlayedInSectors(playerID, wt.getPartlyTrainingSectors(), walkoverWin));
+                        tp.addOsmosisTrainingMinutes(mlt.getTrainingMinutesPlayedInSectors(playerID, wt.getOsmosisTrainingSectors(), walkoverWin));
+                    }
+                    tp.addPlayedMinutes(mlt.getTrainingMinutesPlayedInSectors(playerID, null, walkoverWin));
+                    output.addExperienceIncrease(min(90,tp.getPlayedMinutes() - minutes), type );
+                    minutes = tp.getPlayedMinutes();
+                }
+                TrainingPoints trp = new TrainingPoints(wt, tp);
+
+                // get experience increase of national matches
+                if  ( this.getNationalTeamID() != 0 && this.getNationalTeamID() != myID){
+                    // TODO check if national matches are stored in database
+                    var nationalMatches = train.getMatches(this.getNationalTeamID());
+                    for (var match : nationalMatches){
+                        MatchLineupTeam mlt = DBManager.instance().getMatchLineupTeam(SourceSystem.HATTRICK.getValue(), match.getMatchID(), this.getNationalTeamID());
+                        minutes = mlt.getTrainingMinutesPlayedInSectors(playerID, null, false);
+                        if ( minutes > 0 ) {
+                            output.addExperienceIncrease(min(90,minutes), mlt.getMatchType());
+                        }
+                    }
+                }
+
+                if (TrainingManager.TRAININGDEBUG) {
+                    HOLogger.instance().debug(getClass(), "Week " + train.getHattrickDate().getWeek()
+                            +": Player " + this.getFullName() + " (" + playerID + ")"
+                            +" played total " + tp.getPlayedMinutes() + " mins for training purposes and got "
+                            + wt.getPrimaryTraining(tp) + " primary training points and "
+                            + wt.getSecondaryTraining(tp) + " secondary training points");
+                }
+                output.setTrainingPair(trp);
+                output.setTrainingWeek(train);
+            } catch (Exception e) {
+                HOLogger.instance().log(getClass(),e);
+            }
+        }
+        return output;
     }
 
     private void addExperienceSub(double experienceSub) {
